Got correct answer (debugging)

+++ types.hpp
@@ -3,7 +3,7 @@
 
 #include <limits>
 
-typedef long double TNum;
+typedef double TNum;
 
 const TNum EPS = (TNum) std::numeric_limits<TNum>::epsilon() * 10000000.;
 const TNum PI = 3.14159265358979323846;
